pyproject.toml
[project]
name = "modelo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.2",
]
